requirements.txt
roboflow
python-dotenv
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pycocotools>=2.0.6
albumentations>=1.3.0
pandas>=2.0.0